Add resource type authorization

+++ app/Http/Controllers/Admin/ResourceTypeController.php
@@ -3,55 +3,52 @@
 namespace App\Http\Controllers\Admin;
 
 use App\ResourceType;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Redirect;
-use TiMacDonald\Validation\Rule;
+use App\Http\Requests\ResourceTypeRequest;
 
 class ResourceTypeController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', ResourceType::class);
         $resourceTypes = ResourceType::ordered()->get();
         return view('admin.resource_types.index', compact('resourceTypes'));
     }
 
     public function create()
     {
+        $this->authorize('create', ResourceType::class);
         $resourceType = new ResourceType();
         $formAction = route('admin.resource_types.store');
         return view('admin.resource_types.edit', compact('resourceType', 'formAction'));
     }
 
-    public function store(Request $request)
+    public function store(ResourceTypeRequest $request)
     {
-        ResourceType::create($this->validateValues($request));
+        $this->authorize('create', ResourceType::class);
+        ResourceType::create($request->validated());
         return Redirect::route('admin.resource_types.index');
     }
 
     public function edit(ResourceType $resourceType)
     {
+        $this->authorize('view', $resourceType);
         $formAction = route('admin.resource_types.update', ['resource_type' => $resourceType]);
         return view('admin.resource_types.edit', compact('resourceType', 'formAction'));
     }
 
-    public function update(Request $request, ResourceType $resourceType)
+    public function update(ResourceTypeRequest $request, ResourceType $resourceType)
     {
-        $resourceType->update($this->validateValues($request));
+        $this->authorize('update', $resourceType);
+        $resourceType->update($request->validated());
         return Redirect::route('admin.resource_types.index');
     }
 
     public function destroy(ResourceType $resourceType)
     {
+        $this->authorize('delete', $resourceType);
         $resourceType->deleteIfNotReferenced();
         return Redirect::route('admin.resource_types.index');
-    }
-
-    private function validateValues(Request $request)
-    {
-        return $request->validate([
-            'name' => Rule::required()->string(1, ResourceType::DD_NAME_LENGTH)->get(),
-            'category' => Rule::required()->in(array_keys(ResourceType::CATEGORIES))->get()
-        ]);
     }
 }
